+++ android/app/src/main/java/poke/rogue/helper/presentation/biome/detail/BiomePokemonViewHolder.kt
@@ -1,4 +1,4 @@
-package poke.rogue.helper.presentation.biome.detail.wild
+package poke.rogue.helper.presentation.biome.detail
 
 import androidx.recyclerview.widget.RecyclerView
 import poke.rogue.helper.databinding.ItemPokemonListPokemonBinding
@@ -7,7 +7,7 @@ import poke.rogue.helper.presentation.dex.model.PokemonUiModel
 import poke.rogue.helper.presentation.type.view.TypeChip
 import poke.rogue.helper.presentation.util.view.dp
 
-class BiomeWildItemViewHolder(
+class BiomePokemonViewHolder(
     private val binding: ItemPokemonListPokemonBinding
 ) : RecyclerView.ViewHolder(binding.root) {
