+++ android/app/src/main/java/poke/rogue/helper/presentation/biome/detail/boss/BiomeBossViewHolder.kt
@@ -2,7 +2,7 @@ package poke.rogue.helper.presentation.biome.detail.boss
 
 import androidx.recyclerview.widget.RecyclerView
 import poke.rogue.helper.databinding.ItemBiomePokemonBinding
-import poke.rogue.helper.presentation.biome.detail.wild.BiomeWildItemAdapter
+import poke.rogue.helper.presentation.biome.detail.BiomPokemonAdapter
 import poke.rogue.helper.presentation.biome.model.BiomePokemonUiModel
 import poke.rogue.helper.presentation.util.view.GridSpacingItemDecoration
 import poke.rogue.helper.presentation.util.view.dp
@@ -10,8 +10,7 @@ import poke.rogue.helper.presentation.util.view.dp
 class BiomeBossViewHolder(private val binding: ItemBiomePokemonBinding) :
     RecyclerView.ViewHolder(binding.root) {
 
-        // todo BiomeWildItem이 아니라 BiomePokemonAdapter로 따로 분리할 것
-    private val pokemonAdapter: BiomeWildItemAdapter by lazy { BiomeWildItemAdapter() }
+    private val pokemonAdapter: BiomPokemonAdapter by lazy { BiomPokemonAdapter() }
 
     fun bind(bossPokemon: BiomePokemonUiModel) {
         binding.biomePokemon = bossPokemon
