+++ android/app/src/main/java/poke/rogue/helper/presentation/biome/detail/gym/BiomeGymViewHolder.kt
@@ -2,7 +2,7 @@ package poke.rogue.helper.presentation.biome.detail.gym
 
 import androidx.recyclerview.widget.RecyclerView
 import poke.rogue.helper.databinding.ItemBiomeGymBinding
-import poke.rogue.helper.presentation.biome.detail.wild.BiomeWildItemAdapter
+import poke.rogue.helper.presentation.biome.detail.BiomPokemonAdapter
 import poke.rogue.helper.presentation.biome.model.BiomePokemonUiModel
 import poke.rogue.helper.presentation.util.view.GridSpacingItemDecoration
 import poke.rogue.helper.presentation.util.view.dp
@@ -10,7 +10,7 @@ import poke.rogue.helper.presentation.util.view.dp
 class BiomeGymViewHolder(private val binding: ItemBiomeGymBinding) :
     RecyclerView.ViewHolder(binding.root) {
 
-    private val pokemonAdapter: BiomeWildItemAdapter by lazy { BiomeWildItemAdapter() }
+    private val pokemonAdapter: BiomPokemonAdapter by lazy { BiomPokemonAdapter() }
 
     fun bind(gymPokemon: BiomePokemonUiModel) {
         binding.gymLeader = gymPokemon
